Fix an off-by-one accounting error in the number of duplicated elements
The number of duplicated elements was incremented by 1 on every entry
of duplicate1() and once again by the number of elements whenever an
atomic vector was copied, resulting in an off-by-one error for these
vectors.

Fixed by adjusting the second increment to compensate for the one at
the beginning of the function.