Add experimental peak cons counting
Add an experimental counter for determining the maximum number of at
one time. It is not yet sure if this count is accurate because it uses
another flag bit and it has not yet been verified if there are any
assignments of the full SEXPREC structure in R that would modify this
bit it outside of cons().