Remove counters from DO_CHILDREN in memory.c
Remove the counters from the DO_CHILDREN macro in memory.c and
also the associated outputs. These counters were supposed to count
unused structures, but the DO_CHILDREN macro is the wrong place
to do this as it is used multiple times in the garbage collector
to iterate over all in-memory data structures, so in a single GC
run the same counters could be incremented multiple times for the
same object.